misc: Extract the method to set via properties in a class
