+++ src/controllers/links/Collections.php
@@ -132,12 +132,7 @@ class Collections
 
         if (!auth\LinksAccess::canUpdate($user, $link)) {
             $link = $user->obtainLink($link);
-
-            list($via_type, $via_resource_id) = $this->extractViaFromPath($from);
-            if ($via_type) {
-                $link->via_type = $via_type;
-                $link->via_resource_id = $via_resource_id;
-            }
+            utils\ViaHelper::setLinkVia($link, $from);
         }
 
         $link->is_hidden = $is_hidden;
@@ -155,49 +150,5 @@ class Collections
         }
 
         return Response::found($from);
-    }
-
-    /**
-     * Return the via type and resource id from a path.
-     *
-     * For instance:
-     *
-     * - For the path `/collections/1234567890`, ['collection', '1234567890']
-     *   will be returned (if the collection exists in db)
-     * - For the path `/p/1234567890`, ['user', '1234567890'] will be
-     *   returned (if the user exists in db)
-     * - For other paths, ['', null] will be returned
-     *
-     * @param string $path
-     *
-     * @return string[]
-     */
-    public function extractViaFromPath($path)
-    {
-        $matches = [];
-
-        $result = preg_match('#^/collections/(?P<id>\d+)$#', $path, $matches);
-        if (isset($matches['id'])) {
-            $collection_id = $matches['id'];
-
-            if (!models\Collection::exists($collection_id)) {
-                return ['', null];
-            }
-
-            return ['collection', $collection_id];
-        }
-
-        $result = preg_match('#^/p/(?P<id>\d+)$#', $path, $matches);
-        if (isset($matches['id'])) {
-            $user_id = $matches['id'];
-
-            if (!models\User::exists($user_id)) {
-                return ['', null];
-            }
-
-            return ['user', $user_id];
-        }
-
-        return ['', null];
     }
 }
