+++ tests/controllers/links/CollectionsTest.php
@@ -716,61 +716,4 @@ class CollectionsTest extends \PHPUnit\Framework\TestCase
         $this->assertFlash('error', 'A security verification failed.');
         $this->assertSame(0, models\LinkToCollection::count());
     }
-
-    public function testExtractViaFromPathWithExistingCollection()
-    {
-        $controller = new Collections();
-        $collection_id = $this->create('collection');
-        $path = \Minz\Url::for('collection', ['id' => $collection_id]);
-
-        list($via_type, $via_resource_id) = $controller->extractViaFromPath($path);
-
-        $this->assertSame('collection', $via_type);
-        $this->assertSame($collection_id, $via_resource_id);
-    }
-
-    public function testExtractViaFromPathWithExistingUser()
-    {
-        $controller = new Collections();
-        $user_id = $this->create('user');
-        $path = \Minz\Url::for('profile', ['id' => $user_id]);
-
-        list($via_type, $via_resource_id) = $controller->extractViaFromPath($path);
-
-        $this->assertSame('user', $via_type);
-        $this->assertSame($user_id, $via_resource_id);
-    }
-
-    public function testExtractViaFromPathWithNonExistingCollection()
-    {
-        $controller = new Collections();
-        $path = \Minz\Url::for('collection', ['id' => '12345']);
-
-        list($via_type, $via_resource_id) = $controller->extractViaFromPath($path);
-
-        $this->assertSame('', $via_type);
-        $this->assertNull($via_resource_id);
-    }
-
-    public function testExtractViaFromPathWithNonExistingUser()
-    {
-        $controller = new Collections();
-        $path = \Minz\Url::for('profile', ['id' => '12345']);
-
-        list($via_type, $via_resource_id) = $controller->extractViaFromPath($path);
-
-        $this->assertSame('', $via_type);
-        $this->assertNull($via_resource_id);
-    }
-
-    public function testExtractViaFromPathWithUnsupportedPath()
-    {
-        $controller = new Collections();
-        $path = \Minz\Url::for('bookmarks');
-
-        list($via_type, $via_resource_id) = $controller->extractViaFromPath($path);
-
-        $this->assertSame('', $via_type);
-        $this->assertNull($via_resource_id);
-    }
 }
